added test fo templateEditorTests
